requirements.txt
# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Environment and configuration
python-dotenv>=1.0.0

# Dependencies
pydantic>=2.5.0
pydantic-settings>=2.1.0
requests>=2.31.0
httpx>=0.25.0

# AI SDK
mistralai
